consolidate in logger.go
